+++ classes/class.ilSrUserEnrolmentConfigGUI.php
@@ -47,6 +47,10 @@ class ilSrUserEnrolmentConfigGUI extends ilPluginConfigGUI
         $next_class = self::dic()->ctrl()->getNextClass($this);
 
         switch (strtolower($next_class)) {
+            case strtolower(ExcelImportGUI::class);
+                self::dic()->ctrl()->forwardCommand(new ExcelImportGUI());
+                break;
+
             case strtolower(WorkflowsGUI::class);
                 if (!self::srUserEnrolment()->enrolmentWorkflow()->hasAccess(self::dic()->user()->getId(), false)) {
                     die();
@@ -132,7 +136,7 @@ class ilSrUserEnrolmentConfigGUI extends ilPluginConfigGUI
             return;
         }
 
-        ilUtil::sendSuccess(self::plugin()->translate("saved", self::LANG_MODULE), true);
+        ilUtil::sendSuccess(self::plugin()->translate("configuration_saved", self::LANG_MODULE), true);
 
         self::dic()->ctrl()->redirect($this, self::CMD_CONFIGURE);
     }